tools/external_services/testplan_service/pyproject.toml
[project]
name = "testplan-service"
version = "1.0.0"
description = "3GPP Test Plan Generation Service"
authors = [
    {name = "TestPlan Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0", 
    "python-docx>=0.8.11",
    "duckdb>=0.9.0",
    "polars>=0.20.0",
    "pyarrow>=10.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["services", "models", "core"]